+++ src/main/java/umc/th/juinjang/api/pencil/service/PencilCommandService.java
@@ -1,14 +1,12 @@
 package umc.th.juinjang.api.pencil.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
 import java.util.Optional;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import umc.th.juinjang.api.apple.service.AppleService;
 import umc.th.juinjang.api.apple.service.command.AppleTransactionVerifyCommand;
 import umc.th.juinjang.api.pencil.controller.request.AppleIAPPurchaseRequest;
@@ -27,163 +25,162 @@ import umc.th.juinjang.event.publisher.PaymentEventPublisher;
 @RequiredArgsConstructor
 public class PencilCommandService {
 
-	private final AppleService appleService;
+    private final AppleService appleService;
+    private final PurchasedPencilUpdater purchasedPencilUpdater;
+    private final PurchasedPencilFinder purchasedPencilFinder;
+    private final AcquiredPencilFinder acquiredPencilFinder;
+    private final PencilAccountFinder pencilAccountFinder;
+    private final PaymentEventPublisher paymentEventPublisher;
 
-	private final PurchasedPencilUpdater purchasedPencilUpdater;
-	private final PurchasedPencilFinder purchasedPencilFinder;
-	private final AcquiredPencilFinder acquiredPencilFinder;
-	private final PencilAccountFinder pencilAccountFinder;
-	private final PaymentEventPublisher paymentEventPublisher;
+    @Transactional
+    public Boolean markAcquiredPencilAsRead(Long acquiredPencilId) {
+        AcquiredPencil acquiredPencil = acquiredPencilFinder.findById(acquiredPencilId);
 
-	@Transactional
-	public Boolean markAcquiredPencilAsRead(Long acquiredPencilId) {
-		AcquiredPencil acquiredPencil = acquiredPencilFinder.findById(acquiredPencilId);
+        if (acquiredPencil == null) {
+            throw new EntityNotFoundException("AcquiredPencil not found with id: " + acquiredPencilId);
+        }
 
-		if (acquiredPencil == null) {
-			throw new EntityNotFoundException("AcquiredPencil not found with id: " + acquiredPencilId);
-		}
+        acquiredPencil.updateIsReadAsTrue();
+        return true;
+    }
 
-		acquiredPencil.updateIsReadAsTrue();
-		return true;
-	}
+    @Transactional
+    public AppleIAPPurchaseResponse processAppleIAPPurchase(AppleIAPPurchaseRequest request, Member member,
+            LocalDateTime now) {
+        String transactionId = request.getTransactionId();
+        Long purchaseQuantity = request.getPencilQuantity();
+        Optional<PurchasedPencil> existing = purchasedPencilFinder.findByTransactionIdAndMember(transactionId, member);
 
-	@Transactional
-	public AppleIAPPurchaseResponse processAppleIAPPurchase(AppleIAPPurchaseRequest request, Member member,
-		LocalDateTime now) {
-		String transactionId = request.getTransactionId();
-		Long purchaseQuantity = request.getPencilQuantity();
-		Optional<PurchasedPencil> existing = purchasedPencilFinder.findByTransactionIdAndMember(transactionId, member);
+        if (existing.isEmpty()) {
+            // 기존 트랜잭션이 없는 경우
+            log.info("기존 트랜잭션이 없습니다. {}", transactionId);
+            return validateAndCommitApplePurchase(request, member, now);
+        }
 
-		if (existing.isEmpty()) {
-			// 기존 트랜잭션이 없는 경우
-			log.info("기존 트랜잭션이 없습니다. {}", transactionId);
-			return validateAndCommitApplePurchase(request, member, now);
-		}
+        PurchasedPencil pencil = existing.get();
+        TransactionStatus status = pencil.getTransactionStatus();
+        PencilAccount buyer = pencilAccountFinder.findByMember(member);
 
-		PurchasedPencil pencil = existing.get();
-		TransactionStatus status = pencil.getTransactionStatus();
-		PencilAccount buyer = pencilAccountFinder.findByMember(member);
+        if (status == TransactionStatus.SUCCESS) {
+            // 트랜잭션이 정상적으로 성공된 기록이 있는 경우
+            return AppleIAPPurchaseResponse.ofSuccess(transactionId, purchaseQuantity, buyer.getTotalBalance());
+        }
 
-		if (status == TransactionStatus.SUCCESS) {
-			// 트랜잭션이 정상적으로 성공된 기록이 있는 경우
-			return AppleIAPPurchaseResponse.ofSuccess(transactionId, purchaseQuantity, buyer.getTotalBalance());
-		}
+        PurchasedPencil newPencil = retryPurchasedPencil(request, pencil, member); // 실패 재시도 처리
+        return AppleIAPPurchaseResponse.of(transactionId, newPencil.getTransactionStatus(), purchaseQuantity,
+                buyer.getTotalBalance());
+    }
 
-		PurchasedPencil newPencil = retryPurchasedPencil(request, pencil, member); // 실패 재시도 처리
-		return AppleIAPPurchaseResponse.of(transactionId, newPencil.getTransactionStatus(), purchaseQuantity,
-			buyer.getTotalBalance());
-	}
+    @Transactional
+    public AppleIAPPurchaseResponse validateAndCommitApplePurchase(AppleIAPPurchaseRequest request, Member member,
+            LocalDateTime now) {
+        String transactionId = request.getTransactionId();
 
-	@Transactional
-	public AppleIAPPurchaseResponse validateAndCommitApplePurchase(AppleIAPPurchaseRequest request, Member member,
-		LocalDateTime now) {
-		String transactionId = request.getTransactionId();
+        VerificationResult verificationResult = appleService.verifyAppleTransaction(
+                AppleTransactionVerifyCommand.fromRequest(request));
 
-		VerificationResult verificationResult = appleService.verifyAppleTransaction(
-			AppleTransactionVerifyCommand.fromRequest(request));
+        if (VerificationResult.isSuccess(verificationResult)) {
+            // 성공 시, DB에 저장
+            handleSuccessfulApplePurchase(request, member, now);
 
-		if (VerificationResult.isSuccess(verificationResult)) {
-			// 성공 시, DB에 저장
-			handleSuccessfulApplePurchase(request, member, now);
+            paymentEventPublisher.publishPaymentEvent(member, request.getPrice(), request.getPencilQuantity(),
+                    TransactionStatus.SUCCESS);
+            PencilAccount buyer = pencilAccountFinder.findByMember(member);
+            return AppleIAPPurchaseResponse.ofSuccess(transactionId, request.getPencilQuantity(),
+                    buyer.getTotalBalance());
+        } else {
+            // 실패 시, DB에 저장
+            handleFailureApplePurchase(request, member, now);
 
-			paymentEventPublisher.publishPaymentEvent(member, request.getPrice(), request.getPencilQuantity(),
-				TransactionStatus.SUCCESS);
-			PencilAccount buyer = pencilAccountFinder.findByMember(member);
-			return AppleIAPPurchaseResponse.ofSuccess(transactionId, request.getPencilQuantity(),
-				buyer.getTotalBalance());
-		} else {
-			// 실패 시, DB에 저장
-			handleFailureApplePurchase(request, member, now);
+            paymentEventPublisher.publishPaymentEvent(member, request.getPrice(), request.getPencilQuantity(),
+                    TransactionStatus.VALIDATION_FAILED);
+            return AppleIAPPurchaseResponse.ofValidationFailure(transactionId);
+        }
+    }
 
-			paymentEventPublisher.publishPaymentEvent(member, request.getPrice(), request.getPencilQuantity(),
-				TransactionStatus.VALIDATION_FAILED);
-			return AppleIAPPurchaseResponse.ofValidationFailure(transactionId);
-		}
-	}
+    @Transactional
+    public void handleSuccessfulApplePurchase(AppleIAPPurchaseRequest request, Member member, LocalDateTime now) {
+        String transactionId = request.getTransactionId();
+        Long pencilAmount = request.getPencilQuantity();
 
-	@Transactional
-	public void handleSuccessfulApplePurchase(AppleIAPPurchaseRequest request, Member member, LocalDateTime now) {
-		String transactionId = request.getTransactionId();
-		Long pencilAmount = request.getPencilQuantity();
+        String title = createTitle(pencilAmount);
+        PencilAccount buyer = pencilAccountFinder.findByMemberWithLock(member);
+        buyer.increasePurchasedBalance(pencilAmount);
 
-		String title = createTitle(pencilAmount);
-		PencilAccount buyer = pencilAccountFinder.findByMemberWithLock(member);
-		buyer.increasePurchasedBalance(pencilAmount);
+        purchasedPencilUpdater.save(
+                PurchasedPencil.successOf(member, title, pencilAmount, buyer.getTotalBalance(), request.getPrice(),
+                        request.getPlayTime(), transactionId, request.getAppAccountToken(), now));
 
-		purchasedPencilUpdater.save(
-			PurchasedPencil.successOf(member, title, pencilAmount, buyer.getTotalBalance(), request.getPrice(),
-				request.getPlayTime(), transactionId, request.getAppAccountToken(), now));
+    }
 
-	}
+    @Transactional
+    public void handleFailureApplePurchase(AppleIAPPurchaseRequest request, Member member, LocalDateTime now) {
+        String transactionId = request.getTransactionId();
+        Long pencilAmount = request.getPencilQuantity();
 
-	@Transactional
-	public void handleFailureApplePurchase(AppleIAPPurchaseRequest request, Member member, LocalDateTime now) {
-		String transactionId = request.getTransactionId();
-		Long pencilAmount = request.getPencilQuantity();
+        String title = createTitle(pencilAmount);
+        purchasedPencilUpdater.save(
+                PurchasedPencil.failedDueToValidation(member, title, pencilAmount, request.getPrice(),
+                        request.getPlayTime(), transactionId, request.getAppAccountToken(), now));
+    }
 
-		String title = createTitle(pencilAmount);
-		purchasedPencilUpdater.save(
-			PurchasedPencil.failedDueToValidation(member, title, pencilAmount, request.getPrice(),
-				request.getPlayTime(), transactionId, request.getAppAccountToken(), now));
-	}
+    @Transactional
+    public PurchasedPencil retryPurchasedPencil(AppleIAPPurchaseRequest request, PurchasedPencil pencil,
+            Member member) {
+        if (pencil.getRetryCount() >= 3) { // 재시도 횟수가 3회 이상일 경우 실패로 처리
+            return pencil;
+        }
 
-	@Transactional
-	public PurchasedPencil retryPurchasedPencil(AppleIAPPurchaseRequest request, PurchasedPencil pencil,
-		Member member) {
-		if (pencil.getRetryCount() >= 3) { // 재시도 횟수가 3회 이상일 경우 실패로 처리
-			return pencil;
-		}
+        VerificationResult verificationResult = appleService.verifyAppleTransaction(
+                AppleTransactionVerifyCommand.fromRequest(request)
+        );
 
-		VerificationResult verificationResult = appleService.verifyAppleTransaction(
-			AppleTransactionVerifyCommand.fromRequest(request)
-		);
+        if (VerificationResult.isSuccess(verificationResult)) {
+            pencil.markAsSuccess();
+            pencil.updateRetryCount(pencil.getRetryCount() + 1);
 
-		if (VerificationResult.isSuccess(verificationResult)) {
-			pencil.markAsSuccess();
-			pencil.updateRetryCount(pencil.getRetryCount() + 1);
+            pencilAccountFinder.findByMemberWithLock(member)
+                    .increasePurchasedBalance(pencil.getPurchaseQuantity());
+        }
 
-			pencilAccountFinder.findByMemberWithLock(member)
-				.increasePurchasedBalance(pencil.getPurchaseQuantity());
-		}
+        return pencil;
+    }
 
-		return pencil;
-	}
+    private String createTitle(Long pencilAmount) {
+        return String.format("연필 %d개 구매", pencilAmount);
+    }
 
-	private String createTitle(Long pencilAmount) {
-		return String.format("연필 %d개 구매", pencilAmount);
-	}
+    @Transactional
+    public void handleRefundPurchase(String transactionId) {
+        PurchasedPencil pencil = purchasedPencilFinder.findByTransactionId(transactionId)
+                .orElseThrow(
+                        () -> new EntityNotFoundException("PurchasedPencil not found with transactionId: " + transactionId));
 
-	@Transactional
-	public void handleRefundPurchase(String transactionId) {
-		PurchasedPencil pencil = purchasedPencilFinder.findByTransactionId(transactionId)
-			.orElseThrow(
-				() -> new EntityNotFoundException("PurchasedPencil not found with transactionId: " + transactionId));
+        log.info("Refund processed for transactionId: {}", transactionId);
 
-		log.info("Refund processed for transactionId: {}", transactionId);
+        pencil.markAsRefund();
 
-		pencil.markAsRefund();
+        PencilAccount buyerAccount = pencilAccountFinder.findByMemberWithLock(pencil.getMember());
+        executeRefund(buyerAccount, pencil.getPurchaseQuantity(), pencil.getPrice());
 
-		PencilAccount buyerAccount = pencilAccountFinder.findByMemberWithLock(pencil.getMember());
-		executeRefund(buyerAccount, pencil.getPurchaseQuantity(), pencil.getPrice());
+        paymentEventPublisher.publishPaymentEvent(pencil.getMember(), pencil.getPrice(), pencil.getPurchaseQuantity(),
+                TransactionStatus.REFUNDED);
+    }
 
-		paymentEventPublisher.publishPaymentEvent(pencil.getMember(), pencil.getPrice(), pencil.getPurchaseQuantity(),
-			TransactionStatus.REFUNDED);
-	}
+    public void executeRefund(PencilAccount buyerAccount, long pencilQuantity, long price) {
+        long purchasedToUse = Math.min(buyerAccount.getPurchasedBalance(), pencilQuantity);
+        buyerAccount.decreasePurchasedBalance(purchasedToUse);
 
-	public void executeRefund(PencilAccount buyerAccount, long pencilQuantity, long price) {
-		long purchasedToUse = Math.min(buyerAccount.getPurchasedBalance(), pencilQuantity);
-		buyerAccount.decreasePurchasedBalance(purchasedToUse);
+        long remaining = pencilQuantity - purchasedToUse;
+        long acquiredToUse = Math.min(buyerAccount.getAcquiredBalance(), remaining);
+        buyerAccount.decreaseAcquiredBalance(acquiredToUse);
 
-		long remaining = pencilQuantity - purchasedToUse;
-		long acquiredToUse = Math.min(buyerAccount.getAcquiredBalance(), remaining);
-		buyerAccount.decreaseAcquiredBalance(acquiredToUse);
-
-		buyerAccount.increaseTotalRefundAmount(price);
-		// 남은 수량이 0이 아닐 경우 로그 기록
-		if (remaining - acquiredToUse > 0) {
-			log.warn("Not enough balance to fully refund {} pencils. Refunded only {}.", pencilQuantity,
-				(purchasedToUse + acquiredToUse));
-		}
-	}
+        buyerAccount.increaseTotalRefundAmount(price);
+        // 남은 수량이 0이 아닐 경우 로그 기록
+        if (remaining - acquiredToUse > 0) {
+            log.warn("Not enough balance to fully refund {} pencils. Refunded only {}.", pencilQuantity,
+                    (purchasedToUse + acquiredToUse));
+        }
+    }
 
 }
